refactor: optimize comment, and add BackupSecretPhraseButton component (#1)
